# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      err_code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  HTTP500InternalServerError:
    properties:
      err_code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.CCVMCloudInitCreateRequest:
    properties:
      sn_dns:
        description: service DNS for ccvm
        example: 8.8.8.8
        type: string
      sn_gw:
        description: service gateway for ccvm
        example: 10.10.0.1
        type: string
      sn_ip:
        description: service IP for ccvm
        example: 10.10.14.151
        type: string
      sn_nic:
        description: service NIC for ccvm
        example: enp0s21
        type: string
      sn_prefix:
        description: service prefix for ccvm
        example: 16
        type: integer
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.GenCloudInitISOInfo:
    properties:
      atime:
        example: "2026-05-21T10:00:00+09:00"
        type: string
      ctime:
        example: "2026-05-21T10:00:00+09:00"
        type: string
      filepath:
        example: /var/lib/libvirt/images/scvm-cloudinit.iso
        type: string
      mtime:
        example: "2026-05-21T10:00:00+09:00"
        type: string
      size:
        example: 4096
        type: integer
      tmpdir:
        example: /tmp/gencloudinit-123456
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.GenCloudInitResponse:
    properties:
      action:
        example: generate
        type: string
      code:
        example: 200
        type: integer
      message:
        example: ok
        type: string
      val:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.GenCloudInitISOInfo'
    type: object
...
info:
  contact:
    email: ycyun@ablecloud.io
    name: API Support
    url: https://www.ablecloud.io/support
  description: This is a Cube-API server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ablecloud.io/
  title: Cube API
  version: "1.0"
paths:
...
  /cube/cloudinit/ccvm/generate:
    post:
      consumes:
      - application/json
      description: cluster.json과 선택 입력된 service network 정보를 기준으로 CCVM cloud-init ISO를
        생성하고 대상 ablecube host에 복사합니다.
      parameters:
      - description: ccvm cloud-init generate request
        in: body
        name: body
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMCloudInitCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.GenCloudInitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Create CCVM Cloud-Init ISO
      tags:
      - CUBE - CloudInit
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter the token with the Bearer prefix, e.g. Bearer eyJ...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"